BE/requirements.txt
# Backend Dependencies for Architecture Generation System

# ==============================================================================
# CORE DEPENDENCIES
# ==============================================================================

# Web Framework
Flask==2.3.0
flask-cors==4.0.0

# LLM Integration (Azure OpenAI)
azure-openai==0.13.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# ==============================================================================
# OPTIONAL DEPENDENCIES (for future features)
# ==============================================================================

# For switching LLM providers (optional)
# openai==0.27.0
# anthropic==0.7.0

# For YAML configuration (optional)
# pyyaml==6.0

# For JSON schema validation (optional)
# jsonschema==4.17.0

# ==============================================================================
# DEVELOPMENT DEPENDENCIES
# ==============================================================================

# Testing
pytest==7.3.0
pytest-cov==4.1.0

# Code Quality
black==23.3.0
flake8==6.0.0
mypy==1.2.0
